feat:채팅방 작성,삭제 구현

+++ src/main/java/com/team11/shareoffice/chat/service/ChatService.java
@@ -8,6 +8,7 @@ import com.team11.shareoffice.chat.repository.ChatMessageRepository;
 import com.team11.shareoffice.chat.repository.ChatRoomRepository;
 import com.team11.shareoffice.global.dto.ResponseDto;
 import com.team11.shareoffice.global.exception.CustomException;
+import com.team11.shareoffice.global.security.UserDetailsImpl;
 import com.team11.shareoffice.global.util.ErrorCode;
 import com.team11.shareoffice.member.entity.Member;
 import com.team11.shareoffice.member.repository.MemberRepository;
@@ -17,12 +18,14 @@ import com.team11.shareoffice.post.repository.PostRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.messaging.simp.SimpMessageSendingOperations;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.ZoneId;
 import java.time.ZonedDateTime;
 import java.time.format.DateTimeFormatter;
+import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -56,17 +59,28 @@ public class ChatService {
     public void saveMessage(ChatDto message) {
         Member member = memberRepository.findByNickname(message.getSender()).orElseThrow(() -> new CustomException(ErrorCode.INVALID_MEMBER));
         ChatRoom room = chatRoomRepository.findById(message.getRoomId()).orElseThrow(() -> new CustomException(ErrorCode.CHATROOM_NOT_FOUND));
-        ChatMessage chatMessage = new ChatMessage(member, message.getContent(), room);
+        ChatMessage chatMessage = new ChatMessage(member, message.getMessage(), room);
         chatMessageRepository.saveAndFlush(chatMessage);
-        ChatResponseDto responseDto = new ChatResponseDto(chatMessage.getRoom().getId(), chatMessage.getSender().getNickname(), chatMessage.getContent());
+        ChatResponseDto responseDto = new ChatResponseDto(chatMessage.getRoom().getId(), chatMessage.getSender().getNickname(), chatMessage.getMessage());
         template.convertAndSend("/sub/chat/room/" + message.getRoomId(), responseDto);
     }
 
     @Transactional
-    public ResponseDto deleteRoom(Long roomId) {
+    public ResponseDto<?> deleteRoom(Long roomId,Member member) {
         ChatRoom room = chatRoomRepository.findById(roomId).orElseThrow(() -> new CustomException(ErrorCode.CHATROOM_NOT_FOUND));
-        chatRoomRepository.deleteById(room.getId());
+        if (room.getMember().getNickname().equals(member.getNickname()) || room.getOwner().getNickname().equals(member.getNickname())){
+            chatRoomRepository.deleteById(room.getId());
+        }
+        else {
+            throw new CustomException(ErrorCode.INVALID_CHAT_MEMBER);
+        }
         return ResponseDto.setSuccess("채팅방 삭제 성공");
+    }
+
+    @Transactional
+    public ResponseDto<List<?>> getAllChatRooms(Member member){
+//        List<ChatRoom> chatRooms = chatRoomRepository.findAllByMember(member);
+        return ResponseDto.setSuccess(chatRoomRepository.findAllByMember(member));
     }
 
 }
